refactor: convert tsx files to jsx
