+++ src/components/CommunityCard/index.jsx
