+++ src/theme/Navbar/index.jsx
@@ -17,13 +17,6 @@ import styles from './styles.module.css';
 // retrocompatible with v1
 const DefaultNavItemPosition = 'right';
 
-interface ItemProps {
-  to: string;
-  icon: string;
-  label: string;
-  className?: string;
-}
-
 // items defined here instead of config so they can have an associated icon
 var items = [
   {
@@ -43,15 +36,6 @@ var items = [
   //   label: 'Blog',
   // },
 ];
-
-interface NavLinkProps {
-  to: string;
-  href?: string;
-  label: string;
-  activeClassName?: string;
-  prependBaseUrlToHref?: boolean;
-  icon: string;
-}
 
 function NavLink({
   to,
@@ -85,13 +69,7 @@ function NavLink({
   );
 }
 
-interface NavItemProps extends NavLinkProps {
-  items?: ItemProps[];
-  position?: 'right' | 'left';
-  className?: string;
-}
-
-function NavItem({ items, position = DefaultNavItemPosition, className, ...props }: NavItemProps) {
+function NavItem({ items, position = DefaultNavItemPosition, className, ...props }) {
   const navLinkClassNames = (extraClassName, isDropdownItem = false) =>
     clsx(
       {
@@ -139,11 +117,7 @@ function NavItem({ items, position = DefaultNavItemPosition, className, ...props
   );
 }
 
-interface MobileNavItemProps extends NavItemProps {
-  onClick: any;
-}
-
-function MobileNavItem({ items, className, ...props }: MobileNavItemProps) {
+function MobileNavItem({ items, className, ...props }) {
   // Need to destructure position from props so that it doesn't get passed on.
   const navLinkClassNames = (extraClassName, isSubList = false) =>
     clsx(
@@ -327,7 +301,7 @@ function Navbar() {
           <div className="navbar-sidebar__items">
             <div className="menu">
               <ul className="menu__list">
-                {items.map((linkItem: ItemProps, i) => (
+                {items.map((linkItem, i) => (
                   <MobileNavItem {...linkItem} onClick={hideSidebar} key={i} />
                 ))}
               </ul>
